Extract the TriggerCode Groups
